Removed the unused `remove_dups` conversion file.  Added the Douglas fir in, but fixed the minus sign that was causing an error.
Removed whitespace at the end of lines in `Jenkins_biom`.

Added whitespace in `PEcAn_biom`, cleaned up the output element as well, it just kept overwriting the tree_data versions, creating many copies.  This now cleans this up, but fails to return the `pecan_vars` (at present) because I need to run it briefly to check to make sure that's going to work.