This change is focused on an issue raised by @chrispaciorek over email.  The FIA data generated through the Pecan module was predicated on assignments to PFTs rather than 'PalEON' level species assignments.
Using the conversion table we used the PalEON field rather than the `pft` field in `tree_data` to generate the pseudo-`pft`s.

This requires some changes because of sensitivities in the `allometry` package.  There are a set of taxa with no assignment (`''`) these have to be assigned `None` because of the way pecan writes to file.

The taxa with `/` in their name need to have it substituted with a period, otherwise pecan interprets it as an explicit path in the filename.