Added `plyr` and `raster` functionality to the `control_file.Rmd`.
Also adds the `agg_to_plot` function into the `Rmd`, which is a result of pulling out the `Plot_biom_agg.R` code and putting it into the `agg_to_plot` function.

The requirement for `plyr` is a result of some changes in the `Plot_parameters_agg.R` file.  Did some `lintr`ing as well to the same file.

I think that the `Rasterize_biom.R` file needs to be turned into a function, but haven't done that yet.